added more gui elements and error messages, made ok and cancel buttons fucntional

+++ FlightReservation/src/MainMenu.java
@@ -29,24 +29,40 @@ public class MainMenu {
 					case 1:
 						progressBar.loadBar();
 						String originAirportCode = JOptionPane.showInputDialog(" \n Enter 3 digit airport code for origin (Ex. ATL):");
+						if (originAirportCode == null) { break; }
+						if (originAirportCode.length() == 0) { JOptionPane.showMessageDialog(null,
+								"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+							break; }
 						String destinationAirportCode = JOptionPane.showInputDialog(" \n Enter 3 digit airport code for destination (Ex. MCO):");
+						if (destinationAirportCode == null) { break; }
+						if (destinationAirportCode.length() == 0) { JOptionPane.showMessageDialog(null,
+								"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+							break; }
 
-						ArrayList<Flight> matchedFlights = searchFlights(flightDB, originAirportCode, destinationAirportCode);
-						if (matchedFlights.size() == 0) {
-							System.out.println("No flights available");
-						} else {
-							System.out.println("Available Flights: ");
-							for (int i = 0; i < matchedFlights.size(); i++) {
-								System.out.println("\t" + matchedFlights.get(i).getFlightNumber() + " "
-										+ matchedFlights.get(i).getOrigin().getAirportName() + " -> "
-										+ matchedFlights.get(i).getDestination().getAirportName() + " "
-										+ matchedFlights.get(i).getDepartureTime());
-							}
-						}
+								ArrayList<Flight> matchedFlights = searchFlights(flightDB, originAirportCode, destinationAirportCode);
+								if (matchedFlights.size() == 0) {
+									System.out.println("No flights available");
+								}
+								else {
+									System.out.println("Available Flights: ");
+									for (int i = 0; i < matchedFlights.size(); i++) {
+										System.out.println("\t" + matchedFlights.get(i).getFlightNumber() + " "
+												+ matchedFlights.get(i).getOrigin().getAirportName() + " -> "
+												+ matchedFlights.get(i).getDestination().getAirportName() + " "
+												+ matchedFlights.get(i).getDepartureTime());
+									}
+								}
+
+
 						break;
 					case 2:
 						progressBar.loadBar();
 						String flightInput = JOptionPane.showInputDialog("Enter flight number: ");
+						if (flightInput == null) { break; }
+						if (flightInput.length() == 0) { JOptionPane.showMessageDialog(null,
+								"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+							break; }
+						if(flightInput.length() == 0 ) { break; }
 						int selectedFlightNumber = Integer.parseInt(flightInput);
 						Flight selectedFlight = flightDB.getFlight(selectedFlightNumber);
 						if (selectedFlight == null) {
@@ -56,12 +72,28 @@ public class MainMenu {
 						System.out.println(selectedFlight);
 
 						String seatInput = JOptionPane.showInputDialog("Enter empty seat number: ");
+						if (seatInput == null) { break; }
+						if (seatInput.length() == 0) { JOptionPane.showMessageDialog(null,
+								"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+							break; }
 						int selectedSeatNumber = Integer.parseInt(seatInput);
 						Seat selectedSeat = selectedFlight.getPassengerSeat(selectedSeatNumber);
 						if (selectedSeat != null && selectedSeat.isSeatAvailability()) {
 							String nameEntered = JOptionPane.showInputDialog("Enter name: ");
+							if (nameEntered == null) { break; }
+							if (nameEntered.length() == 0) { JOptionPane.showMessageDialog(null,
+									"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+								break; }
 							String addressEntered = JOptionPane.showInputDialog("Enter address: ");
+							if (addressEntered == null) { break; }
+							if (addressEntered.length() == 0) { JOptionPane.showMessageDialog(null,
+									"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+								break; }
 							String phoneNumberEntered = JOptionPane.showInputDialog("Enter phone number: ");
+							if(phoneNumberEntered == null) { break; }
+							if (phoneNumberEntered.length() == 0) { JOptionPane.showMessageDialog(null,
+									"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+								break; }
 
 							Passenger userPassenger = new Passenger(nameEntered, addressEntered, phoneNumberEntered);
 							System.out.println(userPassenger);
@@ -74,6 +106,9 @@ public class MainMenu {
 					case 3:
 						progressBar.loadBar();
 						flightInput = JOptionPane.showInputDialog("Enter flight number: ");
+						if (flightInput == null) {
+							break;
+						}
 						selectedFlightNumber = Integer.parseInt(flightInput);
 						selectedFlight = flightDB.getFlight(selectedFlightNumber);
 						if (selectedFlight == null) {
@@ -82,17 +117,27 @@ public class MainMenu {
 						}
 						System.out.println(selectedFlight);
 						seatInput = JOptionPane.showInputDialog("Enter occupied seat number: ");
+						if (seatInput == null) { break; }
+						if (seatInput.length() == 0) { JOptionPane.showMessageDialog(null,
+								"Invalid input", "ERROR", JOptionPane.ERROR_MESSAGE);
+							break; }
 						selectedSeatNumber = Integer.parseInt(seatInput);
 						selectedSeat = selectedFlight.getPassengerSeat(selectedSeatNumber);
 						if (selectedSeat.isSeatAvailability()) {
 							System.out.println("No passenger occupies seat number: " + selectedSeatNumber);
 						} else {
+							System.out.println("Passenger removed from seat " + selectedSeatNumber);
 							selectedFlight.deletePassenger(selectedSeatNumber);
 						}
 						break;
 					case 4:
 						progressBar.loadBar();
 						flightInput = JOptionPane.showInputDialog("Enter flight number: ");
+						if (flightInput == null) { break; }
+						if (flightInput.length() == 0) {
+							JOptionPane.showMessageDialog(null, "Invalid input", "ERROR",
+									JOptionPane.ERROR_MESSAGE);
+							break; }
 						selectedFlightNumber = Integer.parseInt(flightInput);
 						selectedFlight = flightDB.getFlight(selectedFlightNumber);
 						if (selectedFlight == null) {
@@ -105,14 +150,18 @@ public class MainMenu {
 						System.out.println("You have exited the program");
 						System.exit(5);
 					default:
-						System.out.println("Input not recognized!!! \n");
+						JOptionPane.showMessageDialog(null, "Please enter the number of one of " +
+								"the given options", "ERROR", JOptionPane.ERROR_MESSAGE);
+						break;
 				}
 
 				}
 			} catch (Exception e) {
 			// TODO: handle exception
-			System.out.println(" Invalid input !!! \n");
-			String errorLog = scan.next();
+			JOptionPane.showMessageDialog(null, "Invalid input", "ERROR",
+					JOptionPane.ERROR_MESSAGE);
+			//System.out.println(" Invalid input !!! \n");
+			//String errorLog = scan.next();
 		}
 
 		}
